+++ package.json
@@ -10,10 +10,12 @@
     "core-js": "^3.8.3",
     "vue": "^2.6.14",
     "vue-inline-svg": "^2.1.3",
+    "vue-router": "^3.5.1",
     "vue-slick-carousel": "^1.0.6"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~5.0.0",
+    "@vue/cli-plugin-router": "~5.0.0",
     "@vue/cli-service": "~5.0.0",
     "vue-template-compiler": "^2.6.14"
   }
