+++ src/main.js
@@ -1,11 +1,13 @@
 import Vue from 'vue'
 import App from './App.vue'
 import InlineSvg from 'vue-inline-svg';
+import router from './router'
 
 Vue.config.productionTip = false
 Vue.component('inline-svg', InlineSvg);
 
 
 new Vue({
-  render: h => h(App),
+  router,
+  render: h => h(App)
 }).$mount('#app')
